cloud-benchmarks: extract config, add RandomizedCloudNativeClusterBenchmark, make the benchmark mixed (reads vs writes)
